Update getlog.pml.php
The expected input of this is int, but it was formerly float and was throwing tons of errors.

+++ inc/getlog.pml.php
@@ -162,7 +162,7 @@ $now = $now->format( 'Y/m/d H:i:s' );
 |
 */
 if ( isset( $_POST['sp'] ) ) {
-    $start_offset = (float)$_POST['sp'] - 1;
+    $start_offset = (int)$_POST['sp'] - 1;
     $start_from   = SEEK_SET;
     $load_more    = true;
 }
